requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
anthropic>=0.40.0
openai>=1.50.0
python-docx>=1.1.0
python-pptx>=1.0.0
pypdf>=4.0.0
docx2pdf>=0.1.8
mistletoe>=1.3.0
